move ConfigMapHistoryService.go to deployment/manifest/configMapAndSecret repo
