+++ api/restHandler/app/pipeline/history/PipelineHistoryRestHandler.go
@@ -18,6 +18,7 @@ package history
 
 import (
 	"fmt"
+	"github.com/devtron-labs/devtron/pkg/deployment/manifest/configMapAndSecret"
 	"github.com/devtron-labs/devtron/pkg/deployment/manifest/deploymentTemplate"
 	"net/http"
 	"strconv"
@@ -46,7 +47,7 @@ type PipelineHistoryRestHandlerImpl struct {
 	enforcer                            casbin.Enforcer
 	strategyHistoryService              history2.PipelineStrategyHistoryService
 	deploymentTemplateHistoryService    deploymentTemplate.DeploymentTemplateHistoryService
-	configMapHistoryService             history2.ConfigMapHistoryService
+	configMapHistoryService             configMapAndSecret.ConfigMapHistoryService
 	prePostCiScriptHistoryService       history2.PrePostCiScriptHistoryService
 	prePostCdScriptHistoryService       history2.PrePostCdScriptHistoryService
 	enforcerUtil                        rbac.EnforcerUtil
@@ -56,7 +57,7 @@ type PipelineHistoryRestHandlerImpl struct {
 func NewPipelineHistoryRestHandlerImpl(logger *zap.SugaredLogger, userAuthService user.UserService,
 	enforcer casbin.Enforcer, strategyHistoryService history2.PipelineStrategyHistoryService,
 	deploymentTemplateHistoryService deploymentTemplate.DeploymentTemplateHistoryService,
-	configMapHistoryService history2.ConfigMapHistoryService,
+	configMapHistoryService configMapAndSecret.ConfigMapHistoryService,
 	prePostCiScriptHistoryService history2.PrePostCiScriptHistoryService,
 	prePostCdScriptHistoryService history2.PrePostCdScriptHistoryService,
 	enforcerUtil rbac.EnforcerUtil,
