+++ pkg/bulkAction/BulkUpdateService.go
@@ -39,13 +39,13 @@ import (
 	repository2 "github.com/devtron-labs/devtron/pkg/cluster/repository"
 	"github.com/devtron-labs/devtron/pkg/deployment/deployedApp"
 	bean5 "github.com/devtron-labs/devtron/pkg/deployment/deployedApp/bean"
+	"github.com/devtron-labs/devtron/pkg/deployment/manifest/configMapAndSecret"
 	"github.com/devtron-labs/devtron/pkg/deployment/manifest/deployedAppMetrics"
 	"github.com/devtron-labs/devtron/pkg/deployment/manifest/deploymentTemplate"
 	"github.com/devtron-labs/devtron/pkg/deployment/manifest/deploymentTemplate/chartRef"
 	bean3 "github.com/devtron-labs/devtron/pkg/deployment/manifest/deploymentTemplate/chartRef/bean"
 	"github.com/devtron-labs/devtron/pkg/eventProcessor/out"
 	"github.com/devtron-labs/devtron/pkg/pipeline"
-	"github.com/devtron-labs/devtron/pkg/pipeline/history"
 	repository4 "github.com/devtron-labs/devtron/pkg/pipeline/history/repository"
 	"github.com/devtron-labs/devtron/pkg/variables"
 	repository5 "github.com/devtron-labs/devtron/pkg/variables/repository"
@@ -86,7 +86,7 @@ type BulkUpdateServiceImpl struct {
 	pipelineRepository               pipelineConfig.PipelineRepository
 	appRepository                    app.AppRepository
 	deploymentTemplateHistoryService deploymentTemplate.DeploymentTemplateHistoryService
-	configMapHistoryService          history.ConfigMapHistoryService
+	configMapHistoryService          configMapAndSecret.ConfigMapHistoryService
 	pipelineBuilder                  pipeline.PipelineBuilder
 	enforcerUtil                     rbac.EnforcerUtil
 	ciHandler                        pipeline.CiHandler
@@ -106,7 +106,7 @@ func NewBulkUpdateServiceImpl(bulkUpdateRepository bulkUpdate.BulkUpdateReposito
 	pipelineRepository pipelineConfig.PipelineRepository,
 	appRepository app.AppRepository,
 	deploymentTemplateHistoryService deploymentTemplate.DeploymentTemplateHistoryService,
-	configMapHistoryService history.ConfigMapHistoryService,
+	configMapHistoryService configMapAndSecret.ConfigMapHistoryService,
 	pipelineBuilder pipeline.PipelineBuilder,
 	enforcerUtil rbac.EnforcerUtil,
 	ciHandler pipeline.CiHandler,
